ready to encode correctly

+++ clir/scripts/create_parallel.py
@@ -9,8 +9,8 @@ def main(lang):
     db.drop_table(f"wiki.{lang}_en_par")
     db.execute_update(f"""
         CREATE TABLE wiki.{lang}_en_par (
-            {lang}_id       int,
-            en_id           int, 
+            {lang}_id       bigint,
+            en_id           bigint, 
             {lang}_title    text,
             en_title        text, 
             {lang}_length   int,
